Support more CSS properties (#13)
* fixed css parsing bug due to misunderstand CSS docs -_-

* added proper value parsing

* changed number to float

* improved compute values

* refactored compute values

* added support for more properties & values

+++ components/css/src/tokenizer/token.rs
@@ -10,12 +10,12 @@ pub enum Token {
     BadUrl,
     Delim(char),
     Number {
-        value: i32,
+        value: f32,
         type_: NumberType,
     },
-    Percentage(i32),
+    Percentage(f32),
     Dimension {
-        value: i32,
+        value: f32,
         type_: NumberType,
         unit: String,
     },
